requirements.txt
requests==2.28.2
SQLAlchemy==2.0.37
python-dotenv
google-cloud-storage
pandas==2.2.3
numpy==2.2.3
pyarrow==19.0.0
gcsfs
